+++ src/components/title/title.test.tsx
@@ -1,0 +1,10 @@
+import * as React from 'react';
+import { render } from '@testing-library/react';
+
+import { Title, TitleProps } from '.';
+
+describe('Title', () => {
+  it('is truthy', () => {
+    expect(Title).toBeTruthy();
+  });
+});